+++ settings_template.py
@@ -4,7 +4,7 @@
 
 
 #  Rocky Version
-VERSION = '0.3.3'
+VERSION = '0.3.4'
 
 # Path to json file containing hostnames, ip addresses and credentials for monitoring
 HOSTS = 'file.json'
@@ -16,9 +16,22 @@ PORTS_PATH = 'path/ports/'
 ROUTER_NAME_SERVERS = ['8.8.8.8', '1.1.1.1']
 
 #  Public Keys
-# These keys allow Rocky to log in to the SRX Router(s) to update the network configuration
-#
+# These keys allow Rocky and Robot to log in to the SRX Router(s) to update the network configuration
+# Rocky RSA
 ROCKY_RSA = ''
+# Robot RSA
+ROBOT_RSA = ''
+
+# Administrator encrypted password (sha256)
+ADMINISTRATOR_ENCRYP_PASS = ''
+
+# api user with limited access
+API_USER_PASS = 'Ap1S3rVIcE!1'
+
+# Radius server
+RADIUS_SERVER_ADDRESS = ''
+# hashed password
+RADIUS_SERVER_SECRET = ''
 
 # Network Device Repository Path
 DEVICE_CONFIG_PATH = 'path/repo/'
@@ -26,11 +39,27 @@ DEVICE_CONFIG_PATH = 'path/repo/'
 #  Managed Cloud Infrastructure
 # One instance of Rocky can manage multiple clouds and each cloud will consist of one or more regions.
 # This configuration describes the cloud infrastructure managed by this instance of Rocky.
+# Monitoring and Provisioning servers needs access to the cloud infrastructure.
+
+# Inbound Access list
+# The management network of this SRXPod is firewalled to block all inbound traffic.
+# Access is allowed to the management network from this list of external Public IPs.
+MAP_ACCESS_LIST = [
+    {
+        'source_address': '',  # 'any' for all ips
+        'destination_address': '',  # 'any' for all ips
+        'port': '',  # valid values 1 , 0-655
+        'protocol': '',  # 'tcp'/'udp'/'any'
+        'description': '',
+    },
+]
+
 
 clouds = [
     # First Cloud ...
     {
         'name': 'First Cloud',
+        'id': '',
 
         # python-cloudcix api settings
         'CLOUDCIX_API_URL': '',
@@ -40,23 +69,111 @@ clouds = [
         'CLOUDCIX_API_VERSION': 2,
         'CLOUDCIX_API_V2_URL': '',
 
-        # Region settings
-        'region': [
-            # First region
+        # COP inbound access list
+        # The management network of this SRXPod is firewalled to block all inbound traffic.
+        # Access is allowed to the management network from this list of external Public IPs.
+        'COP_ACCESS_LIST': [
             {
-                'name': 'First Region',
+                'source_address': '',
+                'destination_address': '',
+                'port': '',
+                'protocol': '',
+                'description': '',
+            },
+        ],
+
+        # Pod settings
+        'pods': [
+            {
+                'name': 'name of the cop',
+                'id': '',
+                'type': 'cop',
 
                 # Network schema
-                'IPv4': [
+                'IPv4_link_subnet': [
+                    #  this is optional, you can leave them as it is.
                     {
-                        'address_range': 'a.b.c.d/e',
-                        'gateway': 'a.b.c.x',
-                    },
-                    {
-                        'address_range': 'a.b.c.d/e',
-                        'gateway': 'a.b.c.x',
+                        'address_range': '',
+                        'gateway': '',
                     },
                 ],
+                'IPv4_pod_subnets': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+                'IPv6_link_subnet': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+                'IPv6_pod_subnets': [
+                    {
+                        'address_range': '',
+                    },
+                ],
+                'IPv4_RFC1918_subnets': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+
+                # VPN Tunnel from Support Center to the SRXPOD for management.
+                'vpns': [],
+
+            },
+
+            # first region
+            {
+                'name': 'region name',
+                'id': '',
+                'type': 'region',
+
+                # Network schema
+                'IPv4_link_subnet': [
+                    #  this is optional, you can leave them as it is.
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+                'IPv4_pod_subnets': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+                'IPv6_link_subnet': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+                'IPv6_pod_subnets': [
+                    {
+                        'address_range': '',
+                    },
+                ],
+                'IPv4_RFC1918_subnets': [
+                    {
+                        'address_range': '',
+                        'gateway': '',
+                    },
+                ],
+
+                # VPN Tunnel from Support Center to the SRXPOD for management.
+                'vpns': [],
             },
 
             # Second region
@@ -64,41 +181,8 @@ clouds = [
                 'name': 'Second Region',
             },
         ],
-
-        # Router Auth details
-        'ROUTER_USER': 'user',
-        'ROUTER_PASSWORD': 'password',
-
-        # SRX settings
-        # SSH Keys allow Robots in each region to configure 'their' local SRX Routers.
-        'ROBOT_RSA': '',
-
-        # Inbound Access list
-        # The management network of this SRXPod is firewalled to block all inbound traffic.
-        # Access is allowed to the management network from this list of external Public IPs.
-        'MGMT_ACCESS_LIST': {
-            #  If this region is managed by an external COP then the IP address ranges of that COP must be allowed in.
-            #  If this region contains a COP then external access is not required and leave as below.
-            'COP_IPV4': '255.255.255.255/32',
-            'COP_IPV6': 'ffff:ffff:ffff:ffff:ffff:ffff:ffff:ffff/128',
-
-            #  This region must be monitored by a instance of Rocky and therefore the IP address
-            'ROCKY': '1',
-        },
-
-        # VPN Tunnel from Support Center to the SRXPOD for management.
-        'vpns': [],
-
-        # A minimum of /29 IPv4 floating gateway subnet for COP servers
-        'IPv4': {
-            'address_range': 'a.b.c.d/e',
-            'gateway': 'a.b.c.x',
-        },
-
-        # A /48 IPv6 routed subnet for Management and Projects
-        'IPv6': {
-            'address_range': 'A:B:C::/48',
-            'link_subnet': 'X:Y:z::/N',  # /126 per region minimum
-        },
     },
+
+    # second cloud
+    {},
 ]